GraphOptInterface

Documentation for GraphOptInterface.

GraphOptInterface.AbstractGraphOptimizer

GraphOptInterface.BipartiteGraph

GraphOptInterface.BipartiteGraph

GraphOptInterface.Edge

GraphOptInterface.Edge

GraphOptInterface.EdgeIndex

GraphOptInterface.EdgeIndexMap

GraphOptInterface.EdgeIndexMap

GraphOptInterface.GraphIndex

GraphOptInterface.HyperEdge

GraphOptInterface.HyperEdge

GraphOptInterface.HyperEdge

GraphOptInterface.HyperGraph

GraphOptInterface.HyperGraph

GraphOptInterface.HyperMap

GraphOptInterface.HyperMap

GraphOptInterface.HyperNode

GraphOptInterface.Node

GraphOptInterface.Node

GraphOptInterface.NodeIndex

GraphOptInterface.OptiGraph

Base.:==

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.reverse

Base.setindex!

Base.setindex!

Base.setindex!

Base.setindex!

GraphOptInterface.add_edge

GraphOptInterface.add_node

GraphOptInterface.add_subgraph

GraphOptInterface.all_edges

GraphOptInterface.all_incident_edges

GraphOptInterface.all_neighbors

GraphOptInterface.all_nodes

GraphOptInterface.all_subgraphs

GraphOptInterface.build_hypergraph_map

GraphOptInterface.children_incident_edges

GraphOptInterface.edge_index

GraphOptInterface.expand

GraphOptInterface.get_edge_variable

GraphOptInterface.get_edges

GraphOptInterface.get_hyperedge

GraphOptInterface.get_hyperedge

GraphOptInterface.get_mapped_edges

GraphOptInterface.get_mapped_edges

GraphOptInterface.get_mapped_nodes

GraphOptInterface.get_mapped_nodes

GraphOptInterface.get_node_variable

GraphOptInterface.get_nodes

GraphOptInterface.get_nodes

GraphOptInterface.get_nodes_to_depth

GraphOptInterface.get_subgraphs

GraphOptInterface.graph_index

GraphOptInterface.identify_edges

GraphOptInterface.identify_nodes

GraphOptInterface.identify_separators

GraphOptInterface.incident_edges

GraphOptInterface.incident_edges

GraphOptInterface.induced_edges

GraphOptInterface.induced_elements

GraphOptInterface.neighborhood

GraphOptInterface.node_index

GraphOptInterface.node_variables

GraphOptInterface.non_parent_incident_edges

GraphOptInterface.non_parent_neighbors

GraphOptInterface.num_all_nodes

GraphOptInterface.num_all_subgraphs

GraphOptInterface.num_edges

GraphOptInterface.num_nodes

GraphOptInterface.num_subgraphs

GraphOptInterface.parent_incident_edges

GraphOptInterface.parent_neighbors

GraphOptInterface.raw_index

GraphOptInterface.raw_index

GraphOptInterface.raw_index

GraphOptInterface.string

GraphOptInterface.string

GraphOptInterface.supports_graph_interface

Graphs.LinAlg.adjacency_matrix

Graphs.LinAlg.adjacency_matrix

Graphs.LinAlg.incidence_matrix

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_vertex!

Graphs.SimpleGraphs.add_vertex!

Graphs.all_neighbors

Graphs.degree

Graphs.edges

Graphs.edges

Graphs.edgetype

Graphs.edgetype

Graphs.has_edge

Graphs.has_edge

Graphs.has_edge

Graphs.has_vertex

Graphs.has_vertex

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.is_directed

Graphs.ne

Graphs.ne

Graphs.nv

Graphs.nv

Graphs.vertices

Graphs.vertices

Graphs.vertices

MathOptInterface.add_variable

GraphOptInterface.AbstractGraphOptimizer — Type

AbstractGraphOptimizer

Abstract supertype for block-structure-exploiting optimizers.

GraphOptInterface.BipartiteGraph — Type

BipartiteGraph

A simple bipartite graph. Contains two vertex sets to enforce bipartite structure.

GraphOptInterface.BipartiteGraph — Method

BipartiteGraph()

Construct an empty BipartiteGraph with no vertices or edges.

GraphOptInterface.Edge — Type

Edge

An edge represents different types of coupling. For instance an Edge{Tuple{Node}} is an edge the couple variables within a single node. An Edge{Tuple{N,Node}} couple variables across one or more nodes.

GraphOptInterface.Edge — Method

Edge(graph_index::GraphIndex, edge_index::EdgeIndex, nodes::NTuple{N,Node})

Construct a new Edge connecting the given tuple of nodes. Initializes an empty MOI model and edge index map for storing coupling constraints.

GraphOptInterface.EdgeIndex — Type

EdgeIndex

A unique identifier for an Edge using an integer value.

GraphOptInterface.EdgeIndexMap — Type

EdgeIndexMap

A bidirectional mapping between edge variables and their corresponding node variables. Contains mappings from edge variables to (Node, MOI.VariableIndex) tuples and vice versa.

GraphOptInterface.EdgeIndexMap — Method

EdgeIndexMap()

Construct an empty EdgeIndexMap with no variable mappings.

GraphOptInterface.GraphIndex — Type

GraphIndex

A unique identifier for an OptiGraph using a Symbol value.

GraphOptInterface.HyperEdge — Type

HyperEdge

A hyperedge connecting multiple hypernodes in a hypergraph. Contains a set of vertices (hypernodes) that the edge connects.

GraphOptInterface.HyperEdge — Method

HyperEdge(t::HyperNode...)

Construct a HyperEdge from a variable number of hypernode arguments.

GraphOptInterface.HyperEdge — Method

HyperEdge(t::Vector{HyperNode})

Construct a HyperEdge from a vector of hypernodes.

GraphOptInterface.HyperGraph — Type

HyperGraph

A very simple hypergraph type. Contains attributes for vertices and hyperedges.

GraphOptInterface.HyperGraph — Method

HyperGraph()

Construct an empty HyperGraph with no vertices or hyperedges.

GraphOptInterface.HyperMap — Type

HyperMap

A mapping from an OptiGraph to a graph view that supports various graph query functions. Currently supports a HyperGraph as the graph view.

GraphOptInterface.HyperMap — Method

HyperMap(graph::OptiGraph, hypergraph::HyperGraph)

Construct a new HyperMap with empty mappings between the given graph and hypergraph.

GraphOptInterface.HyperNode — Type

HyperNode

Type alias for a hypernode, represented as an Int64.

GraphOptInterface.Node — Type

Node

A node represents a set of variables and associated attributes.

GraphOptInterface.Node — Method

Node(graph_index::GraphIndex, node_index::NodeIndex)

Construct a new Node with the given graph and node indices. Initializes an empty MOI model for storing variables and constraints.

GraphOptInterface.NodeIndex — Type

NodeIndex

A unique identifier for a Node using an integer value.

GraphOptInterface.OptiGraph — Type

OptiGraph

A hierarchical graph structure for representing optimization problems with block structure. Contains nodes (representing subproblems), edges (representing coupling constraints), and subgraphs (for hierarchical decomposition).

Fields

index::GraphIndex: Unique identifier for the graph

parent::Union{Nothing,OptiGraph}: Parent graph (if this is a subgraph)

nodes::Vector{Node}: Nodes in this graph

edges::Vector{Edge}: Edges in this graph

subgraphs::Vector{OptiGraph}: Child subgraphs

node_by_index::OrderedDict{NodeIndex,Node}: Lookup map for nodes

edge_by_index::OrderedDict{EdgeIndex,Edge}: Lookup map for edges

subgraph_by_index::OrderedDict{GraphIndex,OptiGraph}: Lookup map for subgraphs

Base.:== — Method

Base.:(==)(h1::HyperEdge, h2::HyperEdge)

Check if two hyperedges are equal by comparing their vertex sets.

Base.getindex — Method

Base.getindex(hypergraph::HyperGraph, edge::HyperEdge)

Get the integer index of a hyperedge in the hypergraph.

Base.getindex — Method

Base.getindex(hypergraph::HyperGraph, node::HyperNode)

Get the index of a hypernode (returns the node itself as the index).

Base.getindex — Method

Base.getindex(hyper_map::HyperMap, edge::Edge)

Get the HyperEdge that corresponds to the given edge in the hyper map.

Base.getindex — Method

Base.getindex(hyper_map::HyperMap, hyperedge::HyperEdge)

Get the Edge that corresponds to the given hyperedge in the hyper map.

Base.getindex — Method

Base.getindex(hyper_map::HyperMap, node::Node)

Get the HyperNode that corresponds to the given node in the hyper map.

Base.getindex — Method

Base.getindex(hyper_map::HyperMap, vertex::HyperNode)

Get the Node that corresponds to the given vertex in the hyper map.

Base.reverse — Method

Base.reverse(e::HyperEdge)

Hyperedges do not support reversal. Throws an error.

Base.setindex! — Method

Base.setindex!(hyper_map::HyperMap, edge::Edge, hyperedge::HyperEdge)

Set the mapping from edge to hyperedge in the hyper map.

Base.setindex! — Method

Base.setindex!(hyper_map::HyperMap, hyperedge::HyperEdge, edge::Edge)

Set the mapping from hyperedge to edge in the hyper map.

Base.setindex! — Method

Base.setindex!(hyper_map::HyperMap, node::Node, vertex::HyperNode)

Set the mapping from node to vertex in the hyper map.

Base.setindex! — Method

Base.setindex!(hyper_map::HyperMap, vertex::HyperNode, node::Node)

Set the mapping from vertex to node in the hyper map.

GraphOptInterface.add_edge — Method

add_edge(graph::OptiGraph, nodes::NTuple{N,Node})::Edge where N

Add an edge to graph.

GraphOptInterface.add_node — Method

add_node(graph::OptiGraph)::Node

Add a node to graph. The index of the node is determined by the central graph.

GraphOptInterface.add_subgraph — Method

add_subgraph(graph::OptiGraph)

Add a new subgraph to graph.

GraphOptInterface.all_edges — Method

all_edges(graph::OptiGraph)

Return all edges in graph including edges from all subgraphs recursively.

GraphOptInterface.all_incident_edges — Method

all_incident_edges(hyper_map::HyperMap, nodes::Vector{Node})::Vector{Edge}

Return all of the edges within the optigraph in hyper_map that are incident to the vector of supplied nodes.

GraphOptInterface.all_neighbors — Method

all_neighbors(hyper_map::HyperMap, nodes::Vector{Node})::Vector{Node}

Return the neighbor nodes within the optigraph in hyper_map to the vector of supplied optigraph nodes.

GraphOptInterface.all_nodes — Method

all_nodes(graph::OptiGraph)

Return all nodes in graph including nodes from all subgraphs recursively.

GraphOptInterface.all_subgraphs — Method

all_subgraphs(graph::OptiGraph)

Return all subgraphs contained in graph recursively.

GraphOptInterface.build_hypergraph_map — Method

build_hypergraph_map(graph::OptiGraph)

Build a HyperMap that maps the given graph to a HyperGraph representation. Creates hypernodes for each node and hyperedges for each edge in the graph. Returns the complete mapping.

GraphOptInterface.children_incident_edges — Method

children_incident_edges(hyper_map::HyperMap, graph::OptiGraph)::Vector{Edge}

Return all of the optigraph edges that are incident to the supplied graph that are strictly child connections.

GraphOptInterface.edge_index — Method

edge_index(edge::Edge)::EdgeIndex

Return the EdgeIndex of the given edge.

GraphOptInterface.expand — Method

expand(g::HyperGraph, nodes::Vector{HyperNode}, distance::Int64)

Expand the given nodes by the specified distance. Returns a tuple of (new_nodes, new_edges) where new_nodes includes all nodes within distance and new_edges are the induced hyperedges on those nodes.

GraphOptInterface.get_edge_variable — Method

get_edge_variable(edge::Edge, node::Node, variable::MOI.VariableIndex)

Get the edge variable index that corresponds to the given node and variable pair.

GraphOptInterface.get_edges — Method

get_edges(graph::OptiGraph)

Return the vector of edges directly contained in graph (not including subgraph edges).

GraphOptInterface.get_hyperedge — Method

get_hyperedge(hypergraph::HyperGraph, edge_index::Int64)

Get the hyperedge corresponding to the given integer index.

GraphOptInterface.get_hyperedge — Method

get_hyperedge(hypergraph::HyperGraph, hypernodes::Set)

Get the hyperedge connecting the given set of hypernodes.

GraphOptInterface.get_mapped_edges — Method

get_mapped_edges(hyper_map::HyperMap, edges::Vector{Edge})

Get the hypergraph HyperEdge elements that correspond to the supplied optigraph edges.

GraphOptInterface.get_mapped_edges — Method

get_mapped_edges(hyper_map::HyperMap, edges::Vector{HyperEdge})

Get the optigraph Edge elements that correspond to the supplied hypergraph edges.

GraphOptInterface.get_mapped_nodes — Method

get_mapped_nodes(hyper_map::HyperMap, nodes::Vector{Node})

Get the hypernode elements that correspond to the supplied optigraph nodes.

GraphOptInterface.get_mapped_nodes — Method

get_mapped_nodes(hyper_map::HyperMap, nodes::Vector{HyperNode})

Get the optigraph Node elements that correspond to the supplied hypergraph nodes.

GraphOptInterface.get_node_variable — Method

get_node_variable(edge::Edge, variable::MOI.VariableIndex)

Get the (Node, MOI.VariableIndex) tuple that corresponds to the given edge variable.

GraphOptInterface.get_nodes — Method

get_nodes(edge::Edge)

Return a vector of all nodes connected by the given edge.

GraphOptInterface.get_nodes — Method

get_nodes(graph::OptiGraph)

Return the vector of nodes directly contained in graph (not including subgraph nodes).

GraphOptInterface.get_nodes_to_depth — Function

get_nodes_to_depth(graph::OptiGraph, depth::Int=0)

Return nodes in graph up to the specified depth of subgraphs. A depth of 0 returns only nodes in the current graph, depth of 1 includes immediate subgraph nodes, etc.

GraphOptInterface.get_subgraphs — Method

get_subgraphs(graph::OptiGraph)

Return the vector of immediate subgraphs contained in graph.

GraphOptInterface.graph_index — Method

graph_index(graph::OptiGraph)

Return the GraphIndex of the given graph.

GraphOptInterface.identify_edges — Method

identify_edges(hypergraph::HyperGraph,partitions::Vector{Vector{HyperNode}})

Identify both induced partition edges and cut edges given a partition of HyperNode vectors.

GraphOptInterface.identify_nodes — Method

identify_nodes(hypergraph::HyperGraph,partitions::Vector{Vector{HyperEdge}})

Identify both induced partition nodes and cut nodes given a partition of HyperEdge vectors.

GraphOptInterface.identify_separators — Method

identify_separators(bgraph::BipartiteGraph, partitions::Vector; cut_selector=Graphs.degree)

Identify separator elements (cut vertices or cut edges) that separate the given partitions. The cut_selector parameter determines how to assign boundary elements to cuts. It can be:

A function (e.g., Graphs.degree) that compares element degrees

:vertex to always assign vertices as separators

:edge to always assign edges as separators

Returns a tuple (partition_elements, cross_elements) where partition_elements are the elements local to each partition and cross_elements are the separator elements.

GraphOptInterface.incident_edges — Method

incident_edges(hypergraph::HyperGraph,hypernode::HyperNode)

Identify the incident hyperedges to a HyperNode.

GraphOptInterface.incident_edges — Method

incident_edges(hypergraph::HyperGraph,hypernodes::Vector{HyperNode})

Identify the incident hyperedges to a vector of HyperNodes.

GraphOptInterface.induced_edges — Method

induced_edges(hypergraph::HyperGraph,hypernodes::Vector{HyperNode})

Identify the induced hyperedges to a vector of HyperNodes.

NOTE: This currently does not support hypergraphs with unconnected nodes

GraphOptInterface.induced_elements — Method

induced_elements(bgraph::BipartiteGraph, partitions::Vector; cut_selector=Graphs.degree)

Get the induced elements for each partition (i.e., elements local to each partition, excluding separators). This is a convenience function that returns only the first element of identify_separators.

Arguments

bgraph::BipartiteGraph: The bipartite graph

partitions::Vector: Vector of partitions

cut_selector=Graphs.degree: Selector for determining cut assignment (function, :vertex, or :edge)

Returns

A vector of vectors, where each inner vector contains the induced elements for that partition.

GraphOptInterface.neighborhood — Method

neighborhood(g::HyperGraph, nodes::Vector{HyperNode}, distance::Int64)

Retrieve the neighborhood within distance of nodes. Returns a vector containing the original vertices and all vertices within the specified distance.

GraphOptInterface.node_index — Method

node_index(node::Node)::NodeIndex

Return the NodeIndex of the given node.

GraphOptInterface.node_variables — Method

node_variables(edge::Edge)::Vector{Tuple{Node,MOI.VariableIndex}}

Return a vector of tuples where each tuple contains the node and variable index associated with each edge variable.

GraphOptInterface.non_parent_incident_edges — Method

non_parent_incident_edges(hyper_map::HyperMap, subgraph::OptiGraph)::Vector{Edge}

Return all of the optigraph edges that are incident to the supplied subgraph that are strictly not parent connections.

GraphOptInterface.non_parent_neighbors — Method

non_parent_neighbors(hyper_map::HyperMap, subgraph::OptiGraph)::Vector{Node}

Return the neighbor nodes in subgraph based on the optigraph in hyper_map that are not in the parent graph of `subgraph.

GraphOptInterface.num_all_nodes — Method

num_all_nodes(graph::OptiGraph)

Return the total number of nodes in graph including all nodes in subgraphs recursively.

GraphOptInterface.num_all_subgraphs — Method

num_all_subgraphs(graph::OptiGraph)

Return the total number of subgraphs in graph recursively.

GraphOptInterface.num_edges — Method

num_edges(graph::OptiGraph)

Return the number of edges directly contained in graph (not including subgraph edges).

GraphOptInterface.num_nodes — Method

num_nodes(graph::OptiGraph)

Return the number of nodes directly contained in graph (not including subgraph nodes).

GraphOptInterface.num_subgraphs — Method

num_subgraphs(graph::OptiGraph)

Return the number of immediate subgraphs in graph.

GraphOptInterface.parent_incident_edges — Method

parent_incident_edges(hyper_map::HyperMap, subgraph::OptiGraph)::Vector{Edge}

Return all of the optigraph edges that are incident to the supplied subgraph that are strictly parent connections.

GraphOptInterface.parent_neighbors — Method

parent_neighbors(hyper_map::HyperMap, subgraph::OptiGraph)::Vector{Node}

Return the neighbor nodes in subgraph based on the optigraph in hyper_map that are only in the parent graph of `subgraph.

GraphOptInterface.raw_index — Method

raw_index(edge::Edge)

Return the raw integer value of the edge's index.

GraphOptInterface.raw_index — Method

raw_index(node::Node)

Return the raw integer value of the node's index.

GraphOptInterface.raw_index — Method

raw_index(graph::OptiGraph)

Return the raw Symbol value of the graph's index.

GraphOptInterface.string — Method

string(edge::HyperEdge)

Return a string representation of the hyperedge showing its vertices.

GraphOptInterface.string — Method

string(graph::HyperGraph)

Return a string representation of the hypergraph showing vertex and edge counts.

GraphOptInterface.supports_graph_interface — Method

supports_graph_interface(optimizer::MOI.AbstractOptimizer)

Check if the given optimizer supports the graph interface. Returns false for standard MOI optimizers and true for AbstractGraphOptimizer subtypes.

Graphs.LinAlg.adjacency_matrix — Method

Graphs.adjacency_matrix(bgraph::BipartiteGraph)

Return the adjacency matrix of the bipartite graph. Rows correspond to vertices in the first set, columns correspond to vertices in the second set.

Graphs.LinAlg.adjacency_matrix — Method

Graphs.adjacency_matrix(hypergraph::HyperGraph)

Obtain the adjacency matrix from hypergraph. Returns a sparse matrix.

Graphs.LinAlg.incidence_matrix — Method

Graphs.incidence_matrix(hypergraph::HyperGraph)

Obtain the incidence matrix representation of hypergraph. Rows correspond to vertices. Columns correspond to hyperedges. Returns a sparse matrix.

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(bgraph::BipartiteGraph, from::Int64, to::Int64)

Add an edge between vertices from and to. Enforces bipartite structure by requiring that the vertices be in different vertex sets.

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(hypergraph::HyperGraph, hypernodes::HyperNode...)

Add a hyperedge connecting the given hypernodes. If the hyperedge already exists, returns the existing hyperedge. Requires at least 2 hypernodes.

Graphs.SimpleGraphs.add_vertex! — Method

Graphs.add_vertex!(bgraph::BipartiteGraph; bipartite=1)

Add a vertex to the bipartite graph. The bipartite keyword argument specifies which vertex set to add the vertex to (1 or 2). Returns the success status.

Graphs.SimpleGraphs.add_vertex! — Method

Graphs.add_vertex!(hypergraph::HyperGraph)

Add a new vertex (hypernode) to the hypergraph. Returns the new hypernode index.

Graphs.all_neighbors — Method

Graphs.all_neighbors(g::HyperGraph, node::HyperNode)

Return all neighbor nodes of the given node in sorted order. Neighbors are nodes that share at least one hyperedge with the given node.

Graphs.degree — Method

Graphs.degree(g::HyperGraph, v::Int)

Return the degree of vertex v (the number of unique neighbors).

Graphs.edges — Method

Graphs.edges(bgraph::BipartiteGraph)

Return an iterator over all edges in the bipartite graph.

Graphs.edges — Method

Graphs.edges(hypergraph::HyperGraph)

Return an iterator over all hyperedges in the hypergraph.

Graphs.edgetype — Method

Graphs.edgetype(bgraph::BipartiteGraph)

Return the edge type for a bipartite graph (SimpleEdge{Int64}).

Graphs.edgetype — Method

Graphs.edgetype(graph::HyperGraph)

Return the edge type for a hypergraph (HyperEdge).

Graphs.has_edge — Method

Graphs.has_edge(bgraph::BipartiteGraph, from::Int64, to::Int64)

Check if the bipartite graph has an edge from from to to.

Graphs.has_edge — Method

Graphs.has_edge(graph::HyperGraph, edge::HyperEdge)

Check if the hypergraph contains the given hyperedge.

Graphs.has_edge — Method

Graphs.has_edge(graph::HyperGraph, hypernodes::Set{HyperNode})

Check if the hypergraph has a hyperedge connecting the given set of hypernodes.

Graphs.has_vertex — Method

Graphs.has_vertex(bgraph::BipartiteGraph, v::Integer)

Check if the bipartite graph contains vertex v.

Graphs.has_vertex — Method

Graphs.has_vertex(graph::HyperGraph, v::Integer)

Check if the hypergraph contains the given vertex.

Graphs.is_directed — Method

Graphs.is_directed(bgraph::BipartiteGraph)

Bipartite graphs are undirected. Always returns false.

Graphs.is_directed — Method

Graphs.is_directed(graph::HyperGraph)

Hypergraphs are undirected. Always returns false.

Graphs.is_directed — Method

Graphs.is_directed(::Type{BipartiteGraph})

Bipartite graphs are undirected. Always returns false.

Graphs.is_directed — Method

Graphs.is_directed(::Type{HyperGraph})

Hypergraphs are undirected. Always returns false.

Graphs.ne — Method

Graphs.ne(bgraph::BipartiteGraph)

Return the number of edges in the bipartite graph.

Graphs.ne — Method

Graphs.ne(graph::HyperGraph)

Return the number of hyperedges in the hypergraph.

Graphs.nv — Method

Graphs.nv(bgraph::BipartiteGraph)

Return the number of vertices in the bipartite graph.

Graphs.nv — Method

Graphs.nv(graph::HyperGraph)

Return the number of vertices (hypernodes) in the hypergraph.

Graphs.vertices — Method

Graphs.vertices(bgraph::BipartiteGraph)

Return a vector of all vertices in the bipartite graph.

Graphs.vertices — Method

Graphs.vertices(hyperedge::HyperEdge)

Return a vector of all vertices connected by the hyperedge.

Graphs.vertices — Method

Graphs.vertices(graph::HyperGraph)

Return a vector of all vertices in the hypergraph.

MathOptInterface.add_variable — Method

MOI.add_variable(edge::Edge, node::Node, variable::MOI.VariableIndex)

Add a variable to the edge that maps to the given variable on the specified node. Returns the edge's variable index for use in edge constraints.